Move comment component to other folder
